Added about-information of DashboardHub
Hi, I came across the website as well as the Readme.md, but no where have you guys written about what Dashboard Hub is. 
yes, there is a youtube video on your site to explain the same, but it just doesn't cut it. 

I wrote a brief description as well as noted down some features (as seen mentioned in the video). I hope this helps in one way or another 

thanks :)